+++ RELEASENOTES.md
@@ -14,6 +14,7 @@
   - ENDOOM text screen on quit
   - Gamma correction
   - Sound velocity option
+  - Option to control filter type for virtual screen scaling
 
 ## Bug fixes:
   - Fix various issues with intermission screen and end-game screen formatting, including use of widescreen assets, text flow, alignment, pillarbox masking on scrolling finales